+++ libyul/backends/evm/AbstractAssembly.h
@@ -118,7 +118,7 @@ public:
 	virtual void markAsInvalid() = 0;
 };
 
-enum class IdentifierContext { LValue, RValue, VariableDeclaration, NonExternal };
+enum class IdentifierContext { LValue, RValue, VariableDeclaration };
 
 /// Object that is used to resolve references and generate code for access to identifiers external
 /// to inline assembly (not used in standalone assembly mode).