+++ libyul/backends/evm/EVMDialect.cpp
@@ -109,21 +109,13 @@ pair<YulString, BuiltinFunctionForEVM> createFunction(
 	return {name, f};
 }
 
-set<YulString> createReservedIdentifiers(langutil::EVMVersion _evmVersion)
+set<YulString> createReservedIdentifiers()
 {
-	// TODO remove this in 0.9.0. We allow creating functions or identifiers in Yul with the name
-	// basefee for VMs before london.
-	auto baseFeeException = [&](evmasm::Instruction _instr) -> bool
-	{
-		return _instr == evmasm::Instruction::BASEFEE && _evmVersion < langutil::EVMVersion::london();
-	};
-
 	set<YulString> reserved;
 	for (auto const& instr: evmasm::c_instructions)
 	{
 		string name = toLower(instr.first);
-		if (!baseFeeException(instr.second))
-			reserved.emplace(name);
+		reserved.emplace(name);
 	}
 	reserved += vector<YulString>{
 		"linkersymbol"_yulstring,
@@ -290,7 +282,7 @@ EVMDialect::EVMDialect(langutil::EVMVersion _evmVersion, bool _objectAccess):
 	m_objectAccess(_objectAccess),
 	m_evmVersion(_evmVersion),
 	m_functions(createBuiltins(_evmVersion, _objectAccess)),
-	m_reserved(createReservedIdentifiers(_evmVersion))
+	m_reserved(createReservedIdentifiers())
 {
 }