+++ test/libsolidity/semanticTests/structs/delete_struct.sol
@@ -2,23 +2,17 @@ contract test {
     struct topStruct {
         nestedStruct nstr;
         uint topValue;
-        mapping (uint => uint) topMapping;
     }
     uint toDelete;
     topStruct str;
     struct nestedStruct {
         uint nestedValue;
-        mapping (uint => bool) nestedMapping;
     }
     constructor() {
         toDelete = 5;
         str.topValue = 1;
-        str.topMapping[0] = 1;
-        str.topMapping[1] = 2;
 
         str.nstr.nestedValue = 2;
-        str.nstr.nestedMapping[0] = true;
-        str.nstr.nestedMapping[1] = false;
         delete str;
         delete toDelete;
     }
@@ -31,12 +25,6 @@ contract test {
     function getNestedValue() public returns(uint nestedValue){
         nestedValue = str.nstr.nestedValue;
     }
-    function getTopMapping(uint index) public returns(uint ret) {
-        ret = str.topMapping[index];
-    }
-    function getNestedMapping(uint index) public returns(bool ret) {
-        return str.nstr.nestedMapping[index];
-    }
 }
 // ====
 // compileViaYul: also
@@ -44,7 +32,3 @@ contract test {
 // getToDelete() -> 0
 // getTopValue() -> 0
 // getNestedValue() -> 0 #mapping values should be the same#
-// getTopMapping(uint256): 0 -> 1
-// getTopMapping(uint256): 1 -> 2
-// getNestedMapping(uint256): 0 -> true
-// getNestedMapping(uint256): 1 -> false