+++ test/libsolidity/syntaxTests/inlineAssembly/function_definition.sol
@@ -1,7 +1,7 @@
 contract C {
   function f() pure public {
     assembly {
-      function f (a, b  , c ) ->    y,x,z {
+      function g (a, b  , c ) ->    y,x,z {
       }
     }
   }