+++ test/libsolidity/syntaxTests/inlineAssembly/invalid/dot_in_fun_param.sol
@@ -1,5 +1,5 @@
 contract C {
-  function f() public pure {
+  function g() public pure {
     assembly {
       function f(a., x.b) -> t.b, b.. {}
     }