+++ test/libsolidity/syntaxTests/inlineAssembly/invalid/dot_in_multi_vardecl.sol
@@ -1,5 +1,5 @@
 contract C {
-  function f() public pure {
+  function g() public pure {
     assembly {
       function f() -> x, y, z {}
       let a., aa.b := f()