+++ test/libsolidity/syntaxTests/inlineAssembly/leave.sol
@@ -1,7 +1,7 @@
 contract C {
   function f() public pure {
     assembly {
-      function f() {
+      function g() {
         leave
       }
     }