+++ test/libsolidity/syntaxTests/inlineAssembly/shadowing/function_name.sol
@@ -1,0 +1,9 @@
+contract C {
+    function f() public pure {
+        assembly {
+            function f() {}
+        }
+    }
+}
+// ----
+// DeclarationError 3859: (75-90): This declaration shadows a declaration outside the inline assembly block.