+++ test/libsolidity/syntaxTests/inlineAssembly/shadowing/local_variable_by_fun_arg.sol
@@ -1,0 +1,10 @@
+contract C {
+    function f() public pure {
+        uint a;
+        assembly {
+            function g(a) {}
+        }
+    }
+}
+// ----
+// DeclarationError 3859: (102-103): This declaration shadows a declaration outside the inline assembly block.