+++ test/libsolidity/syntaxTests/inlineAssembly/string_literal_switch_case.sol
@@ -1,7 +1,7 @@
 contract C {
   function f() public pure {
     assembly {
-      switch codesize()
+      switch calldatasize()
       case "1" {}
       case "2" {}
     }